requirements.txt
PyYAML
psutil
tqdm
argtyped
numpy
torch >= 1.2.0
texar-pytorch
